refactor: ConsentsFollowUp to split out state
